% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-withpkgdown.Rmd.
\name{say_hi}
\alias{say_hi}
\title{Say hi to someone}
\usage{
say_hi(name, exclamation = TRUE)
}
\arguments{
\item{name}{Name of a person}

\item{exclamation}{Whether to include an exclamation mark}
}
\description{
Say hi to someone
}